backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
pydantic==2.9.2
PyJWT==2.9.0

pytest==8.3.3

